Create setup-and-link for `vc dev` to use
