+++ packages/now-cli/src/types.ts
@@ -277,3 +277,8 @@ export interface PaginationOptions {
   count: number;
   next?: number;
 }
+
+export type ProjectLinkResult =
+  | { status: 'linked'; org: Org; project: Project }
+  | { status: 'not_linked'; org: null; project: null }
+  | { status: 'error'; exitCode: number };
